Pull matlab export function back from the grave
I've not reenabled it, but I want to keep it in the source as a
reference.

The planned approach is to export the raw binary data stream to a file
and also export the current configuration to a matlab struct. At that
point, the export function will call an external Octave script that will
first read the configuration, then the raw data, and reformat it into a
matlab struct and resave into a user-defined filename.